requirements.txt
requests
base58
borsh_construct
near-api-py==0.2.1